+++ server/api.go
@@ -498,7 +498,7 @@ func extractClientAddressFromContext(logger *zap.Logger, ctx context.Context) (s
 		clientAddr = peerInfo.Addr.String()
 	}
 
-	return extractClientAddress(logger, clientAddr)
+	return extractClientAddress(logger, clientAddr, ctx, "context")
 }
 
 func extractClientAddressFromRequest(logger *zap.Logger, r *http.Request) (string, string) {
@@ -509,10 +509,10 @@ func extractClientAddressFromRequest(logger *zap.Logger, r *http.Request) (strin
 		clientAddr = r.RemoteAddr
 	}
 
-	return extractClientAddress(logger, clientAddr)
+	return extractClientAddress(logger, clientAddr, r, "request")
 }
 
-func extractClientAddress(logger *zap.Logger, clientAddr string) (string, string) {
+func extractClientAddress(logger *zap.Logger, clientAddr string, source interface{}, sourceType string) (string, string) {
 	var clientIP, clientPort string
 
 	if clientAddr != "" {
@@ -533,6 +533,13 @@ func extractClientAddress(logger *zap.Logger, clientAddr string) (string, string
 			}
 		}
 		// At this point err may still be a non-nil value that's not a *net.AddrError, ignore the address.
+	}
+
+	if clientIP == "" {
+		if r, isRequest := source.(*http.Request); isRequest {
+			source = map[string]interface{}{"headers": r.Header, "remote_addr": r.RemoteAddr}
+		}
+		logger.Warn("cannot extract client address", zap.String("address_source_type", sourceType), zap.Any("address_source", source))
 	}
 
 	return clientIP, clientPort